+++ packages/core/src/index.ts
@@ -1,6 +1,7 @@
 export * from "./commands.js";
 export * from "./domain.js";
 export * from "./ids.js";
+export * from "./land-generator.js";
 export * from "./presets.js";
 export * from "./road-generator.js";
 export * from "./worker-protocol.js";